Revert to defaults for weight/rho that was used for testing
Former-commit-id: 1254a2f1afea80f575314ca299a637b649ae8dff
Former-commit-id: f5dafb1cc2c3ff5614de485f7edc8cf919059b82

+++ examples/cpp/run_all.cpp
@@ -8,8 +8,8 @@
 #include "reader.h"
 #include "timer.h"
 
-//typedef float real_t;
-typedef double real_t;
+typedef float real_t;
+//typedef double real_t;
 
 int main(int argc, char **argv) {
   using namespace std;
